feat: refactor outlet context state for search
