+++ frontend/src/hooks/useCreateVisitData.js
@@ -5,7 +5,7 @@ export default function useCreateSpotData(spotId) {
   // for redirect
   const navigate = useNavigate();
   // logged in user
-  const [userID] = useOutletContext();
+  const { userID } = useOutletContext();
   // not logged in, redirect to login
   if (userID === undefined) {
     navigate("/login");
